requirements.txt
absl-py==2.1.0
accelerate==1.4.0
aiohappyeyeballs==2.4.6
aiohttp==3.11.13
aiosignal==1.3.2
annotated-types==0.7.0
anytree==2.12.1
async-timeout==5.0.1
attrs==25.1.0
beautifulsoup4==4.13.3
bert-score==0.3.13
bioc==2.1
blis==1.2.0
boto3==1.37.5
botocore==1.37.5
catalogue==2.0.10
certifi==2025.1.31
cfgv==3.4.0
chardet==5.2.0
charset-normalizer==3.4.1
click==8.1.8
cloudpathlib==0.20.0
confection==0.1.5
conllu==4.5.3
contourpy==1.3.1
curated-tokenizers==0.0.9
curated-transformers==0.1.1
cycler==0.12.1
cymem==2.0.11
datasets==3.3.2
Deprecated==1.2.18
dill==0.3.8
distlib==0.3.9
docopt==0.6.2
editdistance==0.8.1
en_core_web_sm @ https://github.com/explosion/spacy-models/releases/download/en_core_web_sm-3.8.0/en_core_web_sm-3.8.0-py3-none-any.whl
en_core_web_trf @ https://github.com/explosion/spacy-models/releases/download/en_core_web_trf-3.8.0/en_core_web_trf-3.8.0-py3-none-any.whl
evaluate==0.4.3
filelock==3.17.0
flair==0.15.1
fonttools==4.56.0
frozenlist==1.5.0
fsspec==2024.12.0
ftfy==6.3.1
gdown==5.2.0
grpcio==1.70.0
huggingface-hub==0.29.1
identify==2.6.8
idna==3.10
intervaltree==3.1.0
jieba==0.42.1
Jinja2==3.1.5
jiwer==3.1.0
jmespath==1.0.1
joblib==1.4.2
jsonlines==4.0.0
kiwisolver==1.4.8
langcodes==3.5.0
langdetect==1.0.9
language_data==1.3.0
language_tool_python==2.9.0
lemminflect==0.2.3
lightning-utilities==0.12.0
lru-dict==1.3.0
lxml==5.3.1
marisa-trie==1.2.1
Markdown==3.7
markdown-it-py==3.0.0
MarkupSafe==3.0.2
matplotlib==3.10.1
mdurl==0.1.2
more-itertools==10.6.0
mpld3==0.5.10
mpmath==1.3.0
multidict==6.1.0
multiprocess==0.70.16
murmurhash==1.0.12
networkx==3.3
nlpaug==1.1.11
nltk==3.9.1
nodeenv==1.9.1
num2words==0.5.14
numpy==1.26.4
nvidia-cublas-cu11==11.11.3.6
nvidia-cuda-cupti-cu11==11.8.87
nvidia-cuda-nvrtc-cu11==11.8.89
nvidia-cuda-runtime-cu11==11.8.89
nvidia-cudnn-cu11==9.1.0.70
nvidia-cufft-cu11==10.9.0.58
nvidia-curand-cu11==10.3.0.86
nvidia-cusolver-cu11==11.4.1.48
nvidia-cusparse-cu11==11.7.5.86
nvidia-nccl-cu11==2.21.5
nvidia-nvtx-cu11==11.8.86
OpenHowNet==2.0
packaging==24.2
pandas==2.2.3
path==17.1.0
pillow==11.1.0
pinyin==0.4.0
platformdirs==4.3.6
pptree==3.1
pre_commit==4.1.0
preshed==3.0.9
propcache==0.3.0
protobuf==5.29.3
psutil==7.0.0
py-rouge==1.1
pyarrow==19.0.1
pydantic==2.10.6
pydantic_core==2.27.2
Pygments==2.19.1
pyparsing==3.2.1
PySocks==1.7.1
python-dateutil==2.9.0.post0
pytorch-lightning==2.5.0.post0
pytorch_revgrad==0.2.0
pytz==2025.1
PyYAML==6.0.2
RapidFuzz==3.12.1
regex==2024.11.6
requests==2.32.3
rich==13.9.4
rouge-score==0.1.2
s3transfer==0.11.3
safetensors==0.5.3
scikit-learn==1.6.1
scipy==1.15.2
segtok==1.5.11
sentencepiece==0.2.0
shellingham==1.5.4
six==1.17.0
smart-open==7.1.0
sortedcontainers==2.4.0
soupsieve==2.6
spacy==3.8.4
spacy-curated-transformers==0.3.0
spacy-legacy==3.0.12
spacy-loggers==1.0.5
sqlitedict==2.1.0
srsly==2.5.1
sympy==1.13.1
tabulate==0.9.0
tensorboard==2.19.0
tensorboard-data-server==0.7.2
terminaltables==3.1.10
textattack==0.3.10
thinc==8.3.4
threadpoolctl==3.5.0
tokenizers==0.21.0
toml==0.10.2
torch==2.6.0+cu118
torchmetrics==1.6.1
tqdm==4.67.1
transformer-smaller-training-vocab==0.4.0
transformers==4.49.0
triton==3.2.0
typer==0.15.2
typing_extensions==4.12.2
tzdata==2025.1
urllib3==2.3.0
virtualenv==20.29.2
wasabi==1.1.3
wcwidth==0.2.13
weasel==0.4.1
Werkzeug==3.1.3
Wikipedia-API==0.8.1
word2number==1.1
wrapt==1.17.2
xxhash==3.5.0
yarl==1.18.3
